注册和登录功能基本可用
