+++ .gitignore
@@ -2,3 +2,4 @@
 .idea
 __pycache__
 db.sqlite3
+migrations
