+++ main_site/templates/main_site/error.html
@@ -1,0 +1,14 @@
+{% extends "main_site/base.html" %}
+{% block content %}
+    <div class="container">
+        <div class="card mt-5">
+            <div class="card-header">
+                错误
+            </div>
+            <div class="card-body">
+                <p class="card-text">{{ error }}</p>
+                <a href="" class="btn btn-danger">返回</a>
+            </div>
+        </div>
+    </div>
+{% endblock %}
