+++ main_site/templates/main_site/index.html
@@ -1,5 +1,12 @@
 {% extends "main_site/base.html" %}
 {% block content %}
+    <style>
+        @media (min-width: 1300px) {
+            .container {
+                max-width: 1300px;
+            }
+        }
+    </style>
     <div class="container">
         <div style="overflow:hidden;width:1300px;height:2125px;">
             <div style="margin-left: -55px;margin-top: -230px;margin-right: -30px;">
