+++ user_info/templates/user_info/login.html
@@ -1,0 +1,23 @@
+{% extends "main_site/base.html" %}
+{% block content %}
+    <div class="container col-2">
+        <form class="align-items-end" style="margin-top: 5em;"  method="post">
+            <div class="form-group">
+                {% csrf_token %}
+                <label for="exampleInputEmail1">用户名</label>
+                <input type="text" class="form-control" name="username">
+            </div>
+            <div class="form-group">
+                <label for="exampleInputPassword1">密码</label>
+                <input type="password" class="form-control" name="password">
+            </div>
+            <div class="row">
+                <div class="col-1">
+                    <button type="submit" class="btn btn-danger">登录</button>
+                </div>
+            </div>
+        </form>
+    </div>
+{% endblock %}
+
+
