+++ user_info/urls.py
@@ -18,5 +18,7 @@ from django.conf.urls import url, include
 from . import views
 
 urlpatterns = [
-    url(r'', views.index),
+    url(r'register', views.register, name="register"),
+    url(r'login/', views.auth, name="login"),
+    url(r'^$', views.index,name="info"),
 ]
